% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{linter_ex}
\alias{linter_ex}
\title{Linting for Cyclomatic Complexity}
\usage{
linter_ex(x)
}
\value{
Surprised Message!
}
\description{
Cyclomatic Complexity Example
}
\details{
Demonstrates cyclomatic complexity
}
\examples{
linter_ex("lint")
}
\author{
Yourself
}
